requirements.txt
# API & serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic-settings==2.2.1

# LLM
groq>=0.9.0

# Embeddings (local, free)
sentence-transformers==3.0.1

# Vector DB (local, free)
chromadb==0.5.3

# Document parsing
python-docx==1.1.2

# Text processing
langchain==0.2.6
langchain-community==0.2.6

# Utilities
python-dotenv==1.0.1
tqdm==4.66.4
loguru==0.7.2
tenacity==8.3.0
